validate edit slot fallback ad url
